Large Optimisation
Optimisation by a factor 4.7 : from 680ms to 154ms (with a jackpot at
96ms).
ArrayList at building World was really bad especially because of the
add(index, item) to keep the list sorted.
So it was time for a nice tree. First enormous improve was with the
splendid TreeMultiSet from Google Collections. But looking at the range
of weight of the paths (from 1 to 30 for 50k edges), I had to help me
with that fact so here came a house made data structure AllPath.